+++ uniro_backend/src/test/java/com/softeer5/uniro_backend/building/service/BuildingServiceTest.java
@@ -1,0 +1,4 @@
+package com.softeer5.uniro_backend.building.service;
+
+public class BuildingServiceTest {
+}
