+++ uniro_backend/src/test/java/com/softeer5/uniro_backend/external/MapClientImplTest.java
@@ -1,4 +1,4 @@
-package com.softeer5.uniro_backend.building.client;
+package com.softeer5.uniro_backend.external;
 
 import com.softeer5.uniro_backend.external.MapClientImpl;
 import com.softeer5.uniro_backend.common.exception.custom.ElevationApiException;
